Move date interval utils to services
