+++ domains/rem/src/types.ts
@@ -1,7 +1,5 @@
-import type { RelationalData, IntervalType } from '@eventespresso/services';
+import type { RelationalData } from '@eventespresso/services';
 
 export interface RemDomData {
 	relations?: RelationalData;
 }
-
-export type Intervals = { [key in IntervalType]?: string };
