+++ domains/rem/src/ui/datetimeDetails/useDateFormConfig.ts
@@ -6,8 +6,8 @@ import type { EspressoFormProps } from '@eventespresso/form';
 import { Datetime } from '@eventespresso/edtr-services';
 import { validate } from './formValidation';
 import { DateFormShape } from './types';
-import { DATE_FIELDS_TO_USE, INTERVALS } from '../../constants';
-import { intervalsToOptions } from '../../utils/misc';
+import { DATE_FIELDS_TO_USE } from '../../constants';
+import { intervalsToOptions, DATE_INTERVALS } from '@eventespresso/services';
 
 type DateFormConfig = EspressoFormProps<DateFormShape>;
 
@@ -70,7 +70,7 @@ const useDateFormConfig = (datetime: Datetime, config?: Partial<EspressoFormProp
 						name: 'unit',
 						label: __('Unit'),
 						fieldType: 'select',
-						options: intervalsToOptions(pick(['days', 'hours', 'minutes'], INTERVALS)),
+						options: intervalsToOptions(pick(['days', 'hours', 'minutes'], DATE_INTERVALS)),
 					},
 				],
 			},
