+++ packages/services/src/utilities/date/misc.ts
@@ -1,5 +1,9 @@
 import { pipe } from 'ramda';
+import { __ } from '@wordpress/i18n';
 import { setHours, setMinutes, setSeconds, setYear, setMonth, setDate } from 'date-fns/fp';
+import type { OptionsType } from '@eventespresso/adapters';
+
+import { Intervals } from './types';
 
 /**
  * Sets the time of the date object to zero hour
@@ -18,4 +22,16 @@ export const setDateToToday = (date: Date): Date => {
         setMonth(today.getMonth()),
         setYear(today.getFullYear()
     ))(date);
+};
+
+export const DATE_INTERVALS: Intervals = {
+	months: __('month(s)'),
+	weeks: __('week(s)'),
+	days: __('day(s)'),
+	hours: __('hour(s)'),
+	minutes: __('minute(s)'),
+};
+
+export const intervalsToOptions = (intervals: Intervals): OptionsType => {
+	return Object.entries(intervals).map(([value, label]) => ({ value, label }));
 };
